refactor: rename check_mac_env to check_sys_env and add Windows environment detection
